virt-handler: add protection against migration double-starts
We currently have no mechanism to protect us from executing
`MigrateVirtualMachine()` (which instructs virt-launcher to start
the migration) twice which has unfortunately happened on some
production clusters.

If this occurs virt-launcher detects the double-start and fails the
migration immediately after tearing down the migration proxies,
unfortunately in the logs this appears as a "network error"
(because the source virt-launcher cannot reach the target node
 anymore) even if it's clearly not a network error but rather
a migration abort.

This patch prevents this and adds a unit test.

Signed-off-by: Antonio Cardace <acardace@redhat.com>