add functional test for virtctl ssh
Add tests for the native and local ssh methods. The option '--command'
can simplify the tests because it doesn't necessarly requires the
allocation of a terminal for the native case. For the local ssh methods,
a tty is required because we call virtctl port-forward. For this reason,
we rely on exec.Command and call virtctl binary directly

Signed-off-by: Alice Frosi <afrosi@redhat.com>