examples: Add CirrOS SATA example
This just shows off the ability to define a VM with missing volume
disks. For now we can't change the bus of the default disks that are
generated thus this is a SATA based example using Cirros.

This will likely be replaced in the future with a Flavor and Preferences
based example that keeps the bus as virtio by default.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>