e2e test: add Framework that supports common operations
It will keep a client & clean the cluster for you. It is not necessary anymore to clean
the cluster on `BeforeEach` of all tests if you use the framework because it cleans it
for you. It contains a call to `tests.BeforeTestCleanup()` on its `BeforeEach()`.
It also initializes a virtClient that can be used everywhere: in this way you have not
to retrieve the virtClient because it creates one for you, you can simply use it.
Since it cleans the cluster before each test this ensures that every test runs in the same
cluster conditions.

Signed-off-by: fossedihelm <ffossemo@redhat.com>