e2e test: e2e test use Framework
It is not necessary anymore to clean the cluster on `BeforeEach` of all tests if you use the framework
because it cleans it for you. It contains a call to `tests.BeforeTestCleanup()` on its `BeforeEach()`.
It also initializes a virtClient that can be used everywhere: in this way you have not to retrieve the
virtClient because it creates one for you, you can simply use it. This virtClient is a testClient which
behaves like a normal virtClient but in addition track non-namespaces resources to rollback the changes
after the test execution.
It uses the testClient that rollback every changes to non-namespaces resources.
Since it cleans the cluster before each test this ensures that every test runs in the same cluster conditions.
Only a subset of the tests is affected by changes in order to monitoring issues.

Signed-off-by: fossedihelm <ffossemo@redhat.com>