Add an alert for excessive migrations on a VMI
If a VMI has been successfully migrated more than 12 times over a period of 24 hours, which is considerably higher than normal operation including an upgrade, an alert with a severity of warning is being fired.
Note:
A new VMI that hasn't been migrated yet has no datapoints for its `kubevirt_migrate_vmi_succeeded_total` metric.
When it is first migrated, the metric changes to 1, and the `increase()` function doesn't count this change as an increase in the total migrations for that VMI.
Therefore, for a new VMI the alert is being fired after 13 migrations in the specified time period (24 hours).
If the VMI has been migrated at least once in the past, 12 migrations in 24 hours causes the alert to be fired.

Also, fixing some typos.

Signed-off-by: orenc1 <ocohen@redhat.com>