Add imagePullSecrets field
This MR adds support for imagePullSecrets in KubeVirt CR. This way it's
possible to add secrets which are required to pull images from
private/protected repositories.

The imagePullSecrets is propagagted into following components: virt-handler,
virt-controller and virt-api. Note that definition of imagePullSecrets in
virt-operator must be done manually.

When it comes to virt-launcher, propagating imagePullSecrets field can be a
manageability and security concern - the secret would need to be included in
all the namespaces that the VMI could be started in and user could mount the
secret as volume and read it. Therefore, in the final design, the new container
has been added to virt-handler DaemonSet definition that starts and keeps
virt-launcher image running indefinitely. This happens only when an
imagePullSecrets is defined. With this container the image is pre-fetched to
nodes, so whenever the VMI Pod starts, it can just use the already cache
container image without the need of downloading it and using the secret. Thanks
to this, we avoid security problems but the trade-off a small additional
container per node that takes some CPU and memory resources.

Note: This feature is not compatible with the Always pull policy webhook since
that would force virt-launcher to authenticate to the remote registry on every
start instead of using the cached image.

Signed-off-by: Janusz Marcinkiewicz <januszm@nvidia.com>