Currently we threat host-model of a node as a supported cpuModel because the node contains: `cpu-model.node.kubevirt.io/<host-model-name>: "true"` label in order to run host-model libvirt has to disable some features that the node doesn't have and if we define the cpuModels to be `<host-model-name>` of some node the vmi will fail.
In order to fix this i defined new set of labels (`migration-cpu-model.node.kubevirt.io/<model-name>`) just for migrations with host-model
because in migration we need selector that will allow us to migrate to:
- target node with the same host-model and all the required-features
or
- target node that support the source node's host-model as a supported model that has all the required features

these new labels will be used only when first migrating a vm with host-model cpuModel.
As for the other cases we will use `cpu-model.node.kubevirt.io/<model-name>` always and in these cases we don't care about
the required features because these models are well defined(if node has this label it support all the model's features).

Signed-off-by: bmordeha <bmodeha@redhat.com>