Before creating migration target pod, clean orphan pods
Introduce cleanOrphanPods() function that checks which pods
are migration target pods, but the migration does not exist
anymore. If that's the case, the target pod would hang until
it would be killed after timeout expires. As long as target
pods exist new migrations are blocked, therefore this cleans
them to allow a second migration.

Signed-off-by: Itamar Holder <iholder@redhat.com>