Add memory dump pvc to the vmsnapshot but not to vmrestore
Adding the memory dump to the vmsnapshot is for being able to
export the vmsnapshot including the memory dump.
The memory dump should not be in the vmrestore since it has
no use for it being there, it is not used to restore the VM memory.
When going over the volumebackups we check if the volume if memory
dump volume then it should not be restored.

Signed-off-by: Shelly Kagan <skagan@redhat.com>