Fix flakieness of live migration cancellation func tests
A migration policy is used to limit the migration bandwidth.
Before this commit, a single policy was used for all the tests,
with the exact same name and configurations. This caused flakieness
when running tests in parallel.

Now, a migration policy is created uniquely per-VMI in a way
that supports running tests in parallel.

Signed-off-by: Itamar Holder <iholder@redhat.com>