Makefile: add a lint target
It is very tempting to add code to tests/utils.go, even though it is
already too long to read. Let us try to avoid further growth of this
file by adding this lint target.

This simple/stupid validation is intended to start the discussion. It
should be replaced by a proper go linter that can guide us about the
quality of the code we write.

Signed-off-by: Dan Kenigsberg <danken@redhat.com>