cloud-init: make cloud-init 'instance-id' persistent
Before this patch the VMI uuid was used as the cloud-init `uuid`
(which is then used by cloud-init to build the `instance-id`),
this had the side-effect of letting the cloud-init `instance-id`
change on every VM reboot causing the 'on-first boot' cloud-init
configuration to be executed on every restart and not just once.

This for example would cause the VMI to regenerate SSH keys at
every restart which is not desirable (thus causing a ssh public key
mismatch when connnecting with a client).

This patch fixes this by using the VM UID as uuid which
makes the cloud-init 'instance-id' stable, in case the VMI is not
associated to a VM the VMI's UID will be used instead.

Signed-off-by: Antonio Cardace <acardace@redhat.com>