k8s reporter: Fix virt-launcher pid find logic
Since virt-launcher core code was changed,
and the processes are not suffixed with no-fork
anymore, fix the pgrep logic that find the pid
according the new naming.

Both the virt-launcher monitor and virt-launcher itself
have the same network namespace.
We search for the monitor just because it has
a more unique name, but both would work.

Signed-off-by: Or Shoval <oshoval@redhat.com>