flavors: Reject attempts to update VMI Flavor or Preference
Any update of the FlavorMatcher or PreferenceMatcher for a given
VirtualMachineInstance would previously not be reflected until the
VirtualMachineInstance was recreated. Thus for now any attempt to update
either matcher will be rejected.

Users seeking to change the associated Flavor or Preference of a
VirtualMachineInstance should instead update the underlying
VirtualMachine and restart or recreate their VirtualMachineInstance.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>