libvmi: avoid a needless public constant
Exporting constants from a library is problematic. It makes it easy
for the user code to make assumptions about the type, size and content
of the constant, and as such make it harder for the library to change it
in the future. It is better to hide the constant inside the library and
expose only the bare minimum - the functions that make use of it.

With this patch, we reduce the API of libvmi and simplify the signature
of RandName.

Signed-off-by: Dan Kenigsberg <danken@redhat.com>