libvmi: avoid another needless public constant
Users of libvmi should not to know about libvmi.DefaultTestGracePeriod. If
they care about the grace period of their VMs, they can set it with
WithTerminationGracePeriod(). We use WithTerminationGracePeriod(0) to
make shutdown faster in the test. Once we migrate the termination grace
period test to libvmi we would start using this function externally se
we keep it exposed by the package.

Signed-off-by: Dan Kenigsberg <danken@redhat.com>