tests/libvmi: drop compulsory `name` New() argument
Nothing in our code needs to set a name of any libvmi.New() - all
calls use RandName(). Let us simplify the singnature of New() by
dropping the first argument. This simplify user code, and makes it
redundant to expose RandName. If we ever need to set the VMI name, we
can add a WithName() builder option.

Signed-off-by: Dan Kenigsberg <danken@redhat.com>