flavors: Store ControllerRevision containing VirtualMachine{Flavor,Preference}Spec
This change introduces the use of ControllerRevisions to store versioned
copies of VirtualMachine{Flavor,Preference}Spec.

A new VirtualMachine{Flavor,Preference}SpecRevision object has been
introduced that will contain the original APIVersion of the
VirtualMachine{Flavor,Preference, ClusterFlavor, ClusterPreference}
object and a marshaled copy of the associated spec. The new spec
revision object is then itself marshaled into a ControllerRevision.

Once created the name of the ControllerRevision is stored within the
associated {Flavor,Preference}Matcher of the VirtualMachine for future
use. A follow up change will introduce the actual lookup logic that uses
this stored name.

The ControllerRevisions retain a link to the VirtualMachine through an
OwnerReference, so that when the VirtualMachine is deleted the
ControllerRevisions are also removed from the cluster.

TODOs are left within store{Flavor,Preference}Revision to improve our
behaviour when a ControllerRevision already exists. As this will end up
using lookup code being added in a later commit this change is left for
a future follow up.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>