Clean old code (#1994)
HCO keeps check for wrong owner references in the underline operands.
This was done to handle a bug in old release of HCO, and we can assume
that it's no longer relevant.

This PR remove the `doRemoveExistingOwners` method and the `removeExistingOwner`
field from the generic operand, and the `getObjectMeta` function from the
`hcoResourceHooks` interface.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>