Fix BZ-2080547: mis-read the priority class after update (#1995)
* Fix a bug when updating the priority class

Since most of the priority class fiels are emutable, when updating it,
HCO remove and re-create it. Then HCO reads the new created priority
class in order to set its refertence in the HyperConverged
relatedObjects list.

The re-reading often fails with "no found" just because some timing and
cache issues.

This PR is fixing the issue by droping the re-read.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

* fix the functional test in kubevirtci

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>