tests, sriov: Fix finding the iface name by mac
The existing logic may return an empty iface name, when the mac
appears in the VMI status interface report as sourcing from the
domain.

Here is an example of such a report:
```
{
  "mac": "de:ad:00:00:be:ef",
  "name": "sriov",
  "infoSource": "domain"
}
```
(the `interfaceName` parameter is missing)

In such cases, the poll loop should continue waiting for the data
to be updated by the guest-agent or just fail after the timeout.

Signed-off-by: Edward Haas <edwardh@redhat.com>