tests, network, primary-pod: Remove VMI deletion on teardown
The existing testing framework flow requires that the cleanup of the
test be executed at the beginning (setup) of the next test [1].

Therefore, remove all VMI cleanups from the teardown fixtures.

[1] Cleaning up at the teardown step will remove objects before they
are collected by the custom reporting (which helps debug).
This reporting is hooked into ginkgo where it is called only after a
test finishes (either with success or failure).

Signed-off-by: Edward Haas <edwardh@redhat.com>