builder: Include golangci-lint
Introduce golangci-lint [1] into the builder image in order to start
using a linter to detect common mistakes and attempt to place some
order in the code.

The execution of the linter, with its configuration is introduced in a
following commit.

This change also includes the bump of the builder image.

[1] https://github.com/golangci/golangci-lint

Signed-off-by: Edward Haas <edwardh@redhat.com>