virt-controller: Set default network within VM controller
With the introduction of flavors and preferences we now need to ensure
that the default network and associated interface are added to the VMI
*before* applying any preferences. This should ensure any network
interface specific preferences are also applied to the default network
interface ahead of the VirtualMachineInstance being submitted.

To allow this the original `SetDefaultNetworkInterface` code from the
VMI mutation webhook is moved into the ClusterConfig. Allowing the code
to now be shared between the VM controller and the original VMI mutation
webhook caller.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>