vmi_lifecycle_test: consume less memory
The rfe_id:273 test starts a VM with no disk and no operating system.
However, it allocates 256Mi for the VM RAM. This memory size was chosen
as it is the minimal memory required for running a Fedora guest on the
ARM architecture. This motivation is irrelevant for this test and
needlessly causes us to waste 255Mi of runtime RAM.

Let us reduce the memory consumed by this test. This should be repeated
for most test that do not run a guest operating system.

Signed-off-by: Dan Kenigsberg <danken@redhat.com>