ci, reporter, Add step before reporter dump starts
Currently when a test fails a dump is collected from the cluster for
later debugging. This dump includes various ssh expecter sessions to the vmi,
which are also printed to the screen. This makes it hard understanding when
the test ended and when the repoting started.
A simple solution is to add a step before the dump is collected.

Signed-off-by: Ram Lavi <ralavi@redhat.com>