tests: Move k8s reporter to Just[After|Before]Each
To be able to move the test cleanup to AfterEach the k8s reporter that
dump VMs have to run before the VMs get deleted. This change move the
k8s reporter to JustAfterEach so kubevirt resources are retrieved before
delete it.

Also the trace done at k8sreporter.JustBeforeEach is moved to
JustBeforeEach too to follow same pattern.

Signed-off-by: Enrique Llorente <ellorent@redhat.com>