flavors: Remove use of informers and switch to client calls
The use of SharedInformers to watch for and cache both flavors and
preferences has proven problematic. There have been examples in CI of
virt-api and virt-controller workers both encountering issues either
seeing the latest generation of a resource or even seeing the resource
at all during the initial creation of a VirtualMachine.

For the time being we are going to revert back to client calls and might
revisit the use of SharedInformers in the future.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>