Fix "nil pointer exception" (#2011)
The software craches when updating a metric resource, because then we're
trying to send the HyperConverged pointer to a function as an interface.
The function does check for nil, but this is not working, because
interface(nil) != nil.

This PR fixes the bug by not sending the nil interface but an actual
nil, and also changes the check to use reflection in order to make sure
that the function parameter is nil.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>