tests, sriov: Increase the time waiting for GA report on migration
Recent tests run on CI revealed that post migration, the VMI status
is not updated by the guest-agent information even after 30 seconds.

Unfortunately there is also no other way to wait for a sign from the
guest-agent, as the original condition about its connectivity is not
being reset/updated at the moment.

We are left with increasing the time waiting for the information
to reach the VMI status, by waiting longer for the expected mac
address to appear.

Note: Before fixing the helper that waits for the mac address, the
information was collected from the guest itself (thought the console).

Signed-off-by: Edward Haas <edwardh@redhat.com>