Make DomainSpec optional within VirtualMachineInstanceSpec
VirtualMachineInstanceSpec is currently used directly by
VirtualMachineInstance *and* indirectly by VirtualMachine via
VirtualMachineSpec and VirtualMachineInstanceTemplateSpec.

However with the introduction of instancetypes and existing default
device assignment to a VirtualMachineInstance a user no longer needs to
actually provide a DomainSpec to VirtualMachineInstanceTemplateSpec in
order for the VirtualMachine and eventual VirtualMachineInstance it
generates to be valid.

This change makes DomainSpec optional within the schema of
VirtualMachineInstanceSpec allowing DomainSpec to be dropped when
provided as part of a VirtualMachine.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>