Introduce AutoattachInputDevice to Devices
This change introduces a simple switch to enable the auto attachment of
an input device to a VirtualMachineInstance. To retain current behaviour
this defaults to false to ensure Input devices are not added to existing
VirtualMachineInstances.

When enabled a 'default' Input device is added to the
VirtualMachineInstance definition crafted by the VirtualMachine
controller during start up.

The type and bus of the input device can be controlled by a set of
preferences associated with the VirtualMachine, specifically
PreferredInputBus and PreferredInputType. Alternatively a default type
of 'tablet' and bus of 'usb' will be applied by the
VirtualMachineInstance mutation webhook during admission.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>