util: Allow empty Value within PatchOperation
By allowing an empty Value we are now able to test for `null` values.

This does however break the unused remove op that does not expect a
Value field to be present. As such the constant for this op is itself
removed to denote that it is no longer supported by PatchOperation.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>