flavors: Handle existing ControllerRevisions with unexpected data
This change handles a possible corner case when initially storing
ControllerRevisions (CRs) where a CR of a given name might already exist
in our namespace when we attempt to create one. Previously we would
ignore the error raised by our request to create the new CR and assume
that the existing CR was the same as the one we wanted to create.

There *could* be times when this isn't the case so this change now goes
further and checks to see if the marshled data within the CR is what we
expected it to be, if not an error is raised.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>