Fix warning in event emitter (#2014)
* Fix warning in event emmiter

The event emmiter now checks that an interface object parameter is not a
nil interface, but does not check for simple nil. In this case, the
event emmiter writes this warning to the log:

```
Could not construct reference to: '<nil>' due to: 'can't reference a nil
object'. Will not report event...
```

This PR add the simple == nil check back, in addition to the reflection
check.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

* Fix typo: emmitter ==> emitter

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>