Add proxy configurations to the build system when host behind a firewall
make would fail with below error message when build environment behind a firewall.
To cover this configuration, the proxy configurations of the host should be passed
to the docker build environment:

$ export HTTP_PROXY="xxx.com:888"
$ export HTTPS_PROXY="xxx.com:888"
$ export NO_PROXY="127.0.01, .example.com"
$ make

```
Waiting for it to complete...
Another command (pid=16) is running. Waiting for it to complete on the server (server_pid=20)...
Loading: 0 packages loaded
    Fetching com_github_bazelbuild_buildtools; fetching
make: *** [Makefile:187: format] Error 1
```

Signed-off-by: Haibo Xu <haibo1.xu@intel.com>