Let github collapse BUILD.bazel files in certain locations
BUILD.bazel files in pkg/, staging/ and tests/ are 100% auto-generated
and pretty much 100% of the time not something which need a review.

These paths are also the places where most of the code changes happen.

In the only case where they are normally an issue and need to be fixed
is when unintended custom modifications were done. In this case CI will
uncover that.

Signed-off-by: Roman Mohr <rmohr@google.com>