Replace XFAIL with narrow skips to not clutter CI logs
Some of the XFAIL tests are now producing logs for more than a year.
Replace the XFAIL with narrow clear skips. Since xfail seems to be
mostly used to track external issues in some k8s versions, replacing
them with a skip with the implicit feature that it will break when we
introduce new k8s versions in CI which may have this feature fixed.

It is not optimal, since someone completely outside of the project may
want to run tests against their cluster and they may be confused, but it
adds a repeating non-ignorable re-evaluation point while keeping clear
what's supposed to work and what not.

Signed-off-by: Roman Mohr <rmohr@google.com>