Rename flavor API and CRDs to instancetype
The term flavor was taken from the OpenStack Nova project and while it
does encapsulate parts of what the API is trying to achieve within
KubeVirt it isn't entirely obvious to non-OpenStack users. Before the
API and associated CRDs move out of alphav1 this change replaces the use
of Flavor with the more commonly used term of instance type [1].

The use of the term flavor remains within the vm-preset code as this
predates the flavor API and CRD implementation being renamed here. As
highlighted elsewhere this needs to be deprecated and removed in the
future.

RBAC permissions for the Flavor API are retained by the virt-operator in
order for updates from v0.54 to current HEAD and eventually v0.55 to
continue to work. These permissions should then be dropped by the v0.56
release.

[1] https://aws.amazon.com/ec2/instance-types/

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>