virt-launcher: Stop using UID from libvirt metadata
The domain created in the virt-launcher includes special metadata fields
which are persisted in the libvirt domxml.

Persisting data in the domxml and maintaining it is an extra cost which
is better avoided if possible.

This change removes the UID usage from this metadata and moves it to the
domain object representation, under the standard ObjectMeta structure.

The UID set in the domain is expected to be identical to the one set for
the VMI that created it.

Backward compatibility is preserved at the virt-handler side in order to
correctly process old virt-launchers in the system.

Signed-off-by: Edward Haas <edwardh@redhat.com>