goimports: Adjust the local prefix to include the project
The current goimports local prefix covered the kubevirt organization
and not the kubevirt/kubevirt project itself.

It is useful to see in the source files the dependencies which come from
outside the project. Therefore, updating the local prefix allows
the imports to be structured correctly into blocks that exposes local
and non-local dependencies.

Signed-off-by: Edward Haas <edwardh@redhat.com>