virt-launcher: Include the missing UID when listing the domain
When listing the domains in the virt-launcher, one is returned as a
domain object.

This domain object is currently missing the UID (which also corresponds
to the VMI UID).
This is causing a problem when the domains are collected at the
virt-handler cache (at "resyncing virt-launcher domains" stage) and
missing the UID.

The problem was not seen in that past, because the UID had been also
persisted in the libvirt-metadata section (domain spec).
Therefore, for old virt-launchers, the UID will eventually be read from
the libvirt-metadata later in the backward compatibility handling.

Signed-off-by: Edward Haas <edwardh@redhat.com>