virt-operator: Remove explicit subresources permissions
The `get` verb permissions were only required for a short amount of time
to allow updates from v0.43 and can be removed.

The remaining `update` verb permissions are actually already included by
the call to getKubeVirtComponentsRules [1] when generating the
virt-operator ClusterRole.  This call pulls in all rules required by
KubeVirt for the operator to use as can be seen by looking at the
generated rbac-operator.authorization.k8s.yaml.in file [2] from a
previous commit to this PR.

[1] https://github.com/kubevirt/kubevirt/blob/c4b6ae63c5a7f642ab86b0755dabca3b814ecb39/pkg/virt-operator/resource/generate/rbac/operator.go#L339-L340
[2] https://github.com/kubevirt/kubevirt/blob/43aa19c2b5417c9843f6f7ed58125391cd79c6af/manifests/generated/rbac-operator.authorization.k8s.yaml.in#L510-L519

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>