Deprecate VirtualMachineInstancePresets CRD
VirtualMachineInstancePresets are based on the PodPresets k8s resource
and API that injected data into pods at creation time. However this API
never graduated from alpha and was removed in 1.20 [1]. While useful
there are some issues with the implementation that have resulted in
alternative approaches such as  VirtualMachineInstanceTypes and
VirtualMachinePreferences being made available within KubeVirt.

As per the CRD versioning docs this change updates the generated CRD
definition of VirtualMachineInstancePreset marking the currently
available versions of v1 and v1alpha3 as deprecated.

[1] https://github.com/kubernetes/kubernetes/pull/94090
[2] https://kubernetes.io/docs/tasks/extend-kubernetes/custom-resources/custom-resource-definition-versioning/#version-deprecation

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>