virt-launcher: Deprecate MarkedForGracefulShutdown
Stop using the MarkedForGracefulShutdown flag from libvirt metadata.

It has been used to signal virt-hanlder to initiate a graceful shutdown.
Triggered from the virt-launcher, when catching an OS signal.

This change suggests to drop the marking and perform the graceful
shutdown directly. virt-handler will be notified of the action through
the regular domain changes flow.

The field and its usage at the virt-handler side is kept for backward
compatibility with older virt-launchers.

Signed-off-by: Edward Haas <edwardh@redhat.com>