golanfci-lint: Increase timeout
The time it takes for the linters to run seem to have a large
fluctuation range, between 4sec to over 10sec.

As there are no guarantees given for such jobs (e.g. CPU time), we are
left with increasing the timeout period.

This is mainly triggered by the instability (timeouts) on CI runs.
A possible alternative is to define resource requirements for jobs,
reducing the inconsistent run duration results.

Signed-off-by: Edward Haas <edwardh@redhat.com>