e2e: fix test 6987 in operator suite
The test aims to verify that virt-handler is being updated
with the configured maximum amount of devices. The test was
missing the validation of Kubevirt status, whether deployment of new
configuration has finished.

the commit aims to add validation of the operator status indicating
a successful rollout of the component.

Signed-off-by: Igor Bezukh <ibezukh@redhat.com>