passt: Wait for istio proxy container to be "Running" before running passt
When we start a VMI with passt binding with no restricted ports, the passt
proccess is binding to all the free ports.
It measn that if a side car (for exampe a service mesh proxy) needs to bind
to some ports, it has to be done before passt binary starts.

This commit is taking care of the "istio-proxy" to make sure it is running
before handing the resporsibily from the virt-controller to the
virt-handler.

For any other service meshes proxies/sides cars -
A readiness probe can be used to make sure the containers
are ready before starting passt.
Future improvment may be to wait for all the containers to be running in
case of a VMI with passt binding.

Signed-off-by: Alona Paz <alkaplan@redhat.com>