Fix race of deleting vmsnapshot and freeze of vm
When deleting the vm snapshot we firswt update the finalizer
on the content and then delete the content.
The update of the content can retrigger the update updateVMSnapshotContent
which in case the volumesnapshot wasnt created yet will freeze the vm
and if there is still an error with creating the volume snapshot it will
not unfreeze it, then the content will be deleted and the vm will remain
frozen after the snapshot and the content were already deleted.
Instead incase the vmsnapshot is already marked as deleted if the
content should also be deleted then just do nothing, if the content will
not be deleted then just refrain from creating the volume snapshot (and freezing the vm).

Signed-off-by: Shelly Kagan <skagan@redhat.com>