Change the way we remove memory dump association
Instead of updating the phase to dissociating add a Remove flag.
This change is in order to be able to remove the association
even if there is a memory dump in progress - particularly if it
got stuck for some reason.
Needed a new flag to prevent from having a race of two components
updating the phase (previously it could be updated only after
complete/failure which guarantees no other updates)

Signed-off-by: Shelly Kagan <skagan@redhat.com>