passt: exclude istio reserved ports
When a VM with passt interface and no specific ports is created,
passt proccess is binding to all the free ports.
Therefore, if a service mesh proxy has some reserved ports it is using,
it should bind to them before passt is started.

We verify that the istio-proxy container is running before running the
passt proccess.
But the fact the container is running doesn't give us a full guarantee
istio-proxy finshed binding to all the ports it needs.
Therefore, there may be a race and passt will bind to those "reserved"
ports before the istio proxy.
To avoid it, the istio reserved ports are excluded from passt.

In the future we should consider a more generic way to make sure the
service mesh proxy done binding to all the ports it has.
Currently for service meshes other than istio it can be done using a
readiness probe.

Signed-off-by: Alona Paz <alkaplan@redhat.com>