Add functional test covering SEV device management
The test verifies that the SEV device is not advertised in the node's
capacity when the corresponding feature gate is disabled. In order to
successfully execute the test without SEV-capable hardware, a fake
device file is created, so it can be picked by the device manager.

Note: the test is labeled as serial becaue it alters the global kubevirt
configuration.

Signed-off-by: Vasiliy Ulyanov <vulyanov@suse.de>