virt-controller, vmi, Expose multus annotations to vmi
After the creation of the virt-launcher pod, the annotation
k8s.v1.cni.cncf.io/networks-status is added by multus-cni.

This commit copies this annotation to the vmi object if they exist on
the pod annotations.
virt-controller will hold off passing control to virt-hanlder until
networks-status is created on the pod, or a timeout of 1 second has
passed.

Signed-off-by: Ram Lavi <ralavi@redhat.com>