tests: rename libstorage.NewRandomDataVolume*
The renamed functions create a new DataVolume object. The fact that the
name of the DataVolume is chosen randomly is not important to the user
of the functions. We could have implemented them with a monotonously
rising index and no functionality would have been hurt.

Besides, the new names are shorter and easier to read.

Signed-off-by: Dan Kenigsberg <danken@redhat.com>