virt-controller: Apply preferences before and after adding default devices
268e2f142ba0cdfe4f8c297c2cd620bc420d220f previously moved the creation
of a default network interface to the VM controller *before* an
instancetype and set of preferences are applied.

While this then allowed preferences to be applied to the newly created
interface device it did not allow a user to control the overall
behaviour through the `AutoattachPodInterface` `Device` attribute and
associated `PreferredAutoattachPodInterface` preference.

This change now applies device preferences *before* any default volume
disk, input or network interface devices are added. This allows users to
disable or enable the attachment of said devices through a set of
preferences.

The original call to apply the full instancetype and set of preferences
is retained so any devices that are added still have preferences
applied.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>