sriov, map pci-addresses to intefaces with network-status
Currently the pci-addresses are mapped to the SRIOV interfaces using the
SRIOV resource pool [0].
This is not enough to connect when there is more than 1 interfaces from
the same resource pool, or when the interfaces use the same
network-attachement-definition.

This commit utilies the network-status annotation from the vmi to
accurately map the pci-address to each SRIOV interface, while keeping
the legacy mapping method in case the annotation does not exist.

[0]
https://github.com/k8snetworkplumbingwg/sriov-network-device-plugin#configurations

Signed-off-by: Ram Lavi <ralavi@redhat.com>