instancetype: Always use a preference bool value when no user value given
Previously preference booleans would only be applied to a VMI definition
if they were true. This obviously removes the ability for a preference
to disable an attribute that itself defaults to true.

This change reworks the application of preference booleans so that their
value is always used when they are provided and no corresponding user
defined value is already found within the VMI.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>