operator_test: avoid a ton of lambdas
operator_test was using multiple lambdas which incurred unjustified
complexity on the reader. This commit drops the redundant stanza of
function declarations, moves the functions to the scope of the file and
makes minor modifications to pass originalCDI and originalKv as explicit
arguments.

This commit make virClient a file-wide variable. I think that it is
better than passing it again and again to several functions. Still, this
is a hint that there is more to fix here.

Signed-off-by: Dan Kenigsberg <danken@redhat.com>