e2e: increase timeout in operator test 4744
test seems to be flaky, but in the logs it is seen
that the componentes are patched, this is a constant
observation since this flake began. I will increase
the timeout to 90 secs. In case this won't be enough then
I will revert this commit and will observe potential hangs

Signed-off-by: Igor Bezukh <ibezukh@redhat.com>