tests, clientcmd: Move command code to libcli
There is some code reading from flags at clientcmd so it cannot be used
at the k8s reporter if we want to use it as a external lib. This change
create at libcli to move the non flag code from clientcmd there and
depend k8s reporter on it.

Signed-off-by: Enrique Llorente <ellorent@redhat.com>