Update ginkgolinter
Replace the ginkgo-linter from the source code, with a newer version
from https://github.com/nunnatsa/ginkgolinter. This version add two more
validations, in addition to the existing wrong length validation:

* wrong nil assertion; e.g.
   `Expect(x).To(Eaual(nil))` => `Expect(x).To(BeNil()`
   `Expect(x == nil).To(BeTrue())` => `Expect(x).To(BeNil()`
* wrong error assertion; e.g.
   `Expect(err).To(BeNil())` => `Expect(err).ToNot(HaveOccurred())`
   `Expect(funcReturnsError()).To(BeNil())` =>
   `Expect(funcReturnsError()).To(Succeed())`

The reason for taking this linter code out is to ease the development of
it, independatly from the kubevirt development. Also, it will allow
other projects to use it as well.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>